// src/interfaces/User.ts
interface IUser {
  date_of_birth: string;
  date_of_joining: string;
  designation: string;
  email: string;
  first_name: string;
  hired_location: string;
  id: string;
  last_name: string;
  middle_name: string;
  proof_id: string;
  residential_address: string;
  role_id: number;
  salary: number;
  work_status: string;
  years_of_experience: number;
}
export default IUser;